Editorial: stop listing terms defined in other specs
Instead, those terms are defined in a "database" of <dfn> tags that don't make it to the output, and as such don't require so much manual curation. Closes #5919. This should make contributing to HTML much easier.

This also fixes a few cross-links and updates some of the prose in what's left of the dependencies section.