Add WebDriver BiDi navigation callbacks
This change is motivated by WebDriver BiDi needing to observe the
progress of navigations, in order to support commands and events related
to navigation: https://github.com/w3c/webdriver-bidi/pull/93

In HTML, this adds a single new concept; a navigation id which is a
unique id generated for each navigation. This enables WebDriver BiDi to
tell when events are from the same navigation.

In WebDriver BiDi, a navigation status struct is defined, with an id, a
URL and a status. This is used as a uniform interface for communicating
the navigation progress through all the integration points, even though
not all fields are useful in all cases.

The actual integration points added are:

* Navigation started
* Navigation aborted
* Navigation failed
* Download started
* Fragment navigated
* DOM load complete
* Load complete