Editorial: Plugins are no longer an extensibility mechanism
As reported by dg12 on Matrix, with Flash no longer being supported, it
shouldn't be listed as an example of the use of plugins. Furthermore,
since #6003 aims to remove any complexity around plugins except for that
which is needed for PDF viewers, it is clear that plugins in general are
no longer an extensibility mechanism.