Modify pushState()/replaceState()'s URL rewriting check
The actual normative change here is to allow the query component to change in the cross-origin case, not just the fragment component. However, while we're here, we pull out the algorithm into a standalone definition, and rephrase it to be clearer with fewer nested negative conditions.

This better aligns with browsers, exactly matching WebKit and being only slightly more lenient than Chromium and Gecko. See the full analysis at https://github.com/whatwg/html/issues/6836#issuecomment-905821850.

This also improves the domintro for pushState()/replaceState() to be a bit more in line with modern style, and to explain the error condition.

Closes #6836.