Editorial: factor out and rephrase "can have its URL rewritten"
This creates a dedicated algorithm that underlies whether a URL can be rewritten using history.pushState()/replaceState(). The algorithm is phrased differently than the old inline steps, but it has the same normative effect, and is clearer since it involves less negatives.

This also adds a table of examples for "can have its URL rewritten", and modernizes and expands the domintros for the pushState() and replaceState() methods.