Editorial: trivial fixes to COOP algorithms
Firstly, in "process a navigate fetch", responseCOOP is not presently declared, but variables must not be used before they are declared: https://infra.spec.whatwg.org/#variables

Secondly, the "current origin" item of cross-origin opener policy enforcement result is called "origin" at all other uses. The definition is renamed to be consistent with the references, and with the "url" item, which is similarly not qualified as "current".