Explicit algorithm for idle callback "deadline"
The algorithm replaces the algorithm that is currently defined in prose
here: https://w3c.github.io/requestidlecallback/#the-idledeadline-interface).

It works as such:
- Timers can calculate their estimated next callback timestamps,
  based on their start time and document-inactive time so far.
- The deadline for idle tasks for an event loop is the earliest between:
  - The time until the next timer callback in this event loop is estimated
    to fire.
  - The time until the next render, estimated at16ms after the start of
    the previous task that had a rended opportunity, if there are
    pending requestAnimationFrame callbacks.
  - 50ms

TODO: the deadline is currently checked between windows, it needs to be
re-checked between callbacks in the same window.

See https://github.com/w3c/requestidlecallback/issues/71