List "Chinese, Simplified" and "Chinese, Traditional"
For historical reasons going back to Netscape days and predating the introduction
of the Hans and Hant tags, Firefox and Chrome localizations don't use Hans and Hant
to designate the Simplified Chinese and Traditional Chinese localizations.

The previous conclusion of omitting SG, HK, and MO due to the resulting mismatch
with Vista locales was incorrect.