Move Page Visibility APIs into HTML
Defines the document.hidden and document.visibilityState APIs in the HTML Standard, while improving their definitions to be more rigorous. This is similar to 3285b988b8e962d9692e547fd3101ceb23cb1cb6 which did the same for the visibilitychange event.

Also updates the discussion of "page is in the background" to refer to the new definitions.

See https://github.com/w3c/page-visibility/issues/74 and https://github.com/w3c/page-visibility/pull/73 for more background.