Define storage of preloaded resources
A document keeps a list of preloaded resources, each with relevant
parameters from the request, and the response once available.

Once a <link rel=preload> element starts fetching a resource, that
entry is added, and once the response is fully loaded, the fetch
consuming the resource receives the response.

See https://github.com/whatwg/fetch/issues/590