Standardize window.open() popup features and BarProp values
Fixes #5872 by standardizing an internal "is popup" boolean, and how it is impacted by various window.open() features. Implementations can use this boolean (in addition to other signals) to determine whether an opened window is a popup or not.

Fixes #4431 by changing the various BarProp visible properties to only reflect this "is popup" boolean, instead of reflecting the actual state of user interface elements, or the passed-in window features.

Adds a "popup" window.open() feature to provide a simpler method of requesting a popup.