Move away from including IANA templates, part 2
This concludes moving away from using IANA templates to define HTTP headers. It takes care of Last-Event-ID, Ping-From, Ping-To, Refresh, and X-Frame-Options.

It also adds an index to ensure these headers are still easy to find.

And finally it corrects the casing of several index sections.

This helps with #7274 (kept open to update the IANA registry) and fixes #7355.