Remove a `setTimeout`'s entry in the map of active timers after it runs
It used to be unobservable whether a non-repeating timer that had
already run its callback had an entry in the list of active timers or
not, and so those entries were never removed from the list. However,
calculation of the idle deadline, meaning that if a timer expires and is
never cancelled, the deadline will be in the past.

This change avoids this case by removing non-repeating timers from the
map of active timers after their callback runs.

Closes #7376.