Editorial: assert entry list isn't null in the form submission algorithm
The construct the entry list algorithm can return null if the form's
constructing entry list flag is true, but "entry list" is assumed to be
non-null in the rest of the form submission algorithm. This makes sense,
since the constructing entry list flag has been checked in step 2, and
even if the submit event listeners invoked the `FormData` constructor,
the entry list construction would happen synchronously in that listener.

This change adds an assertion to clarify that entry list cannot be null.