Remove concept of "expressly inert"
It seems a remnant of control groups, which were removed in #3647.
All uses are replaced with the concept of "inert".

Also cleaning up "When a key event is to be routed in a top-level
browsing context", which also seems to have remnants of when control
groups allowed inert elements inside an inert dialog to be focusable.

Fixes #7564.