Replace origin with storage key for BroadcastChannel partitioning
This PR updates the BroadcastChannel portion of the HTML standard
to have storage keys be used instead of origins when determining
which browsing contexts should be able to have BroadcastChannel
messages sent and received from them. Today, a storage key only
consists of an origin, but it will likely soon include the
top-level origin as well. Only using origin enables certain forms
of cross-site tracking. For more details, see:

https://privacycg.github.io/storage-partitioning/