Remove the concept of "expressly inert"
It seems to be a remnant of control groups, which were removed in 90a60b2a0dc740b8b0093b07ca0a41e70ba8d83a. All uses are replaced with the concept of "inert".

Also clean up "When a key event is to be routed in a top-level browsing context", which seems to have remnants of when control groups allowed inert elements inside an inert dialog to be focusable.

Fixes #7564.