Update :active and :hover definitions (#7465)
This removes the disabled condition for :active, matching Gecko and WebKit. The condition was added to the spec in 998597345d884fbc6482bf72f66d3a25b11e289d but not implemented uniformly. As discussed in #6635, it doesn't work very well and is inconsistent with other similar pseudo-classes, like :hover.

This also removes the consideration of <link> elements with href="" attributes for :active, since they no longer have activation behavior as of 1e0ee7fc1aa490212a8655ceccaf12accb5b9be9.

Finally, this updates the definitions of both :active and :hover to be based on the latest Selectors spec, which notably includes the ancestor condition over the flat tree at the CSS layer, instead of having HTML handle that. This required splitting up "being activated" and "matches :active", as well as "designated" and "matches :hover".

Closes #6635, although note that there remain other interop issues with :active, tracked in #7578.