Stop propagating inertness into nested browsing contexts
Firefox doesn't propagate it. Blink does, but I'm removing it since it
can be a cross-site leak, and blocks fenced frames since it opens a
communication channel. WebKit only propagates it for `.focus()`, but not
for the selection API.

Note that even if the contents in the nested browsing contexts are not
marked as inert, if the browsing context container is inert, they won't
respond to mouse interactions, and they won't be reached by sequential
navigation.

Fixes #7605.