Stop propagating inertness into nested browsing contexts
This can be a cross-site communications channel, and browsers don't want to leave it open.

Note that even if the contents in the nested browsing contexts are not marked as inert, if the browsing context container is inert, the contents won't respond to mouse interactions, and they won't be reached by sequential focus navigation. So they generally behave similarly from the perspective of the user.

Fixes #7605.