Editorial: modernize the canvas Compositing section
This largely consists of defining new underlying concepts to back getters and setters, and make better use of existing ones.

It also makes the setter algorithm for globalCompositeOperation more precise, by clearly referencing the <composite-mode> and <blend-mode> properties instead of saying "the possible values are those defined in Compositing and Blending". (This did uncover an interop issue which we will track in #7637.)

Finally this renames the "current composition operator" to "current compositing and blending operator" (as per the Compositing and Blending spec's wording) and fixing the uses of "composition" vs "compositing".