Make media.currentTime more accurate during seeking
Previously, currentTime always returned the "official playback position", which is only periodically synced with the "current playback position". In particular, if JavaScript code performs a seek (e.g. using the seek() method), this would not sync them until a stable state is reached. (I.e., until that JavaScript code runs to completion.)

This shows up particularly in the case where currentTime is set to a value outside of the timeline, e.g. Number.MAX_VALUE. The previous spec would cause whatever is passed to the currentTime setter, to then be returned by the currentTime getter. But Chromium and Gecko instead clamped the result to the timeline. Chromium's implementation of this, in particular, is to return the current playback position during the seek process, which is what we have specified here.

See https://github.com/web-platform-tests/wpt/pull/32832 for more discussion.