Set timeOrigin to fetch startTime
No need to measure timestamp during navigation and account for
unload prompts. Fetch already measures the startTime. Patch it through to
`document load info` which is in turn accessed by the
`environment settings object`.

This is in line with current WebKit implementation, and very close to
Chromium implementation which currently doesn't count beforeunload
time as part of the navigation time.

Closes #7716