Be more rigorous about insertion/removal
Part of #2771. This removes the concepts of "node is inserted" and "node is removed" in favor of explicit HTML Standard insertion/removing steps, which are of the form the DOM Standard wants us to use. These HTML Standard insertion/removing steps, in turn, call out to element definition-specific "HTML element insertion/removing steps".

This does not complete #2771, as there remain several less-rigorous definitions, outlined in https://github.com/whatwg/html/issues/2771#issuecomment-1074202744. But it gives us the right scaffolding and prevents any further usage of the node is inserted/removed definitions.