Avoid suggesting the empty string for `navigator.platform`
`navigator.platform` was permitted to return the empty string due to
a privacy-motivated spec change:
https://searchfox.org/whatwg-html/diff/ecefa6e105592656bf592e0abcf0276dad971fbc/source#31141

Since 2008, the understanding of changing things like this
for privacy has evolved from returning special privacy values
to returning a value that is already commonly returned for
non-privacy reasons. Therefore, change the spec not to say
"the empty string" but to permit returning a string commonly
returned on another platform.

Additionally, change the examples to be from the set of values
given at https://www.chromium.org/updates/ua-reduction/ to
avoid suggesting more identifying and potentially less
compatible values, but, for now, avoid actually constraining
the permitted values to a reduced set.