Obsolete the <param> element
Given that plugins are gone from the web platform (with their full removal from the spec being tracked in #6003), it is not useful. In some browsers it can be used to figure out the URL of an <object>, even when that <object> is not being used for a plugin, via params named "movie", "src", "code", or "url". But we decided to remove this behavior from browsers instead of specifying it.

This retains the HTMLParamElement interface, as well as the parser behavior of <param>.

Closes #387. Helps with #6003.