several clarifications
a document is conforming without a h1, but it is not recommended.
a document is conforming without a h1 as the first heading in the DOM, but it is not recommended.
sectioning content does not effect heading levels, but authors are encouraged to match heading level and sectioning content nesting level.
Thanks to @mtrootyy for feedback