Rework media element removal steps
As per #2882 "await a stable state" is a problematic concept as it queues a microtask without encompassing task.

This replaces its usage here by queuing a task as the media element paused member does not change upon element removal directly.

(It's unclear if you can observe the paused member changing before the pause event. As written this seems possible if you are (un)lucky. Potentially this means that the internal pause steps are not correct for this caller. They are correct for at least one other caller: pause().)