Editorial: clarify string attribute reflection and enumerated attributes
Algorithmize USVString, DOMString, and DOMString? reflection. This is easier to read, partially because it spreads out the steps and branches, and partially because of the asserts making some conditions explicit.

Rewrites the section on enumerated attributes to be more clear about how they work. Notably this gives an explicit algorithm for determining the state, instead of a couple paragraphs of prose. It also broadens the definition of "canonical keyword" to automatically be determined in simpler cases, instead of only used in tricky cases.

Finally, make it clear that DOMString?-reflected enumerated attributes (of which crossOrigin/crossorigin=""  is the only one currently) are always "limited to only known values".