Use the custom attributes in a script tag
Refer to Named modules / scripts #7600, proposed to use the custom attributes in a script tag rather to add another new <Name> tag to the Script which would be used as an alias that can be referred to in import statements. This will able to save bandwidth by reduce another new attribute.