Clean events index
This implements the rule suggested in #8340 to have the events table only
include events that the HTML spec actually fires.

Changes made:
- Drop `click`, `contextmenu`, `copy`, `cut`, `input`, `paste`,
`securitypolicyviolation` and `slotchange` entries from the table
- Complete the Dependencies section to mention the `copy`, `cut` and `paste`
events defined in Clipboard API and Events, the `slotchange` event (about to be)
defined in DOM and the `securitypolicyviolation` (about to be) defined in CSP.
- Add Clipboard API and Events to the list of references.
- Drop the reference to Page Visibility and drop Page Visibility from the list
of references.

Pull request should only be merged once DOM and CSP get updated to define the
`securitypolicyviolation` and `slotchange` events.