Fix initiator origin for external protocol navigations
This commit introduces a lightweight non-fetch scheme navigation params, which
holds an initiator origin. It is constructed in two places:
 - Navigations to non-fetch scheme URLs
   - initiator origin is set to document-state-initiator-origin
 - Navigations to fetch scheme URLs that redirect
   - initiator origin is set to the last fetch-scheme URL in the redirect
     chain, which matches what browsers do

One confusing result of this commit is that we now have the following:
 - source-snapshot-params-origin
 - document-state-initiator-origin
 - non-fetch-scheme-params-initiator-origin

I think we should get rid of source-snapshot-params-origin, but I need to see if
it is useful for anything, or if we always want the true *original* navigation
initiator origin for everything.