Fix HostResolveImportedModule's current realm handling
If this is called as part of module script loading of a dependency, there is no current realm. But we don't actually need the current realm, except in the case of a top-level import() inside an event handler, in which case there *will* be a current realm. Rearrange the logic of the algorithm to only look at the current realm (or rather, current settings object) in that later case.

Also removes some vestigial base URL stuff, and camel-cases all the variable names.

Closes #8410.