Fix redirect handling for non-HTTP(s)-scheme URLs
Before this commit, after the redirect-following loop in "create a navigation
params by fetching", we would check if the response was a network error, and if
so, return null. Immediately after, we would check to see if `locationURL` was a
non-fetch-scheme URL, in which case we'd return the appropriate "non-fetch
scheme navigation params" struct.

This was wrong though, and resulted in the non-fetch scheme navigation params
*never* being constructed. That's because for all non-HTTP(s) redirects, the
HTTP-redirect algorithm returns a network error, which means for all external
protocol redirects, we'd satisfy the network error condition and immediately
return null.

This commit fixes things by checking if `locationURL` is a non-fetch-scheme URL
*before* the "is network error" check, since the former is a strict subset of
the latter.

This commit also removes the "locationURL is a URL whose scheme is a fetch
scheme" condition in our return null block, because it is not needed. As
previously mentioned, this condition is covered by the "response is network
error" check, so it is redundant.