Fix what URL is committed to session history for all redirects
Prior to this commit, we set `SHE.url = locationURL` too early; we did this
assignment before we even check if the `locationURL` is a URL that we wouldn't
follow (in particular, non-HTTP(s) URLs). This is inconsistent with the model
that we've proposed in [1] for handling traversals to non-HTTP(s) redirects.

This commit fixes this by moving the step in which we commit `locationURL` to
the SHE to *after* we update the `currentURL` variable, which is only done when
we re-enter the top of the loop to follow the redirect.

[1]: https://docs.google.com/document/d/1yxnzjRDVmAR5CC9GcAyY448lBD0u0E98eUEMHDhx1Dw/edit#heading=h.m78rarim3k4g