Fix the steps to find a point on an ellipse
This is an attempt at fixing #8408.
This goes the long way of defining the steps to determine the point on the ellipse, in a "theoretical" wording (as opposed to using algebra formulae).
This is based on what browsers seems to do.  so there shouldn't be a need for browser issues, yet a technical review from @whatwg/canvas is probably necessary.
Not sure either if editorially this is the proper way of fixing that.