Editorial: simplify session history serialized state
Previously, it could be either a serialized state or null. But, we treated null the same as the serialized form of null (i.e., StructuredSerializeForStorage(null)), when we deserialized it in "restore the history object state". This simplifies the infrastructure always having it be a serialized state, and never null.

Once this is done, it becomes clear that initializing new history entries with their serialized state set to StructuredSerializeForStorage(null) was pointless, since that is the default value. So, we can remove a couple instances of that.