Minor fixes to COOP enforcement result processing
Most notably, this properly creates a COOP enforcement result for the "inline content that doesn't have a DOM" case. Previously, that would just create a new COOP enforcement result struct, but not all of the struct's members have default values, so that was not actually allowed.

Other changes are editorial:

* Use <dl class="props"> whenever constructing new COOP enforcement results. (This was previously done in 2/5 cases.)

* Make use of the default value for "needs a browsing context group switch" and "would need a browsing context group switch due to report-only", instead of always setting them.

* Add a default value for "current context is navigation source".