Report fallback resource timing for no-TAO iframes
When encountering a cross-origin iframe without TAO, don't automatically
report its fetch start/end times, as that might expose user interaction
with the iframe (a cross-origin violation).

Instead, report a fallback resource timing entry, with the navigation start
time and the load event time of the iframe, which are already observable.

Depends on whatwg/fetch#1579
Closes w3c/resource-timing#340