Refactor import-related host hooks to follow the JS spec
Follows https://github.com/tc39/ecma262/pull/2905. See https://docs.google.com/presentation/d/1RVUE-MENQT8dj2wxvMLMDxg_VoMOwiwNQQged39QIEU/edit for an overview of the changes here.

This is not a normative change, and in fact hosts can continue to have the same capabilities as before. But now, much of the graph-crawling work can be offloaded to the JavaScript specification.