Add "check and possibly close popover stack" algorithm
The "auto popover list" is constructed by attributes set on buttons.
When those buttons are modified, it can break connections in the list.
This patch adds checks to spots where buttons can be modified in order
to fix up the list by closing all popovers when a connection has been
broken.