Fix script-blocking style sheet infrastructure
In particular, fixes #4325, which documents a case where <link rel="stylesheet"> would increment, and never decrement, the script-blocking style sheet counter.

To do this without increasing the amount of action-at-a-distance in the spec, we move from a script-blocking style sheet counter to a script-blocking style sheet set. That also matches implementations better, and sets us up for fixing other issues in this area.