Define auto popover list in terms of the top layer
This solves an issue with elements being unconditionally added to this list and makes it clearer that the steps between it being added here and the top layer didn't mean anything.

This also makes hide popover use check popover validity more to rely on the shared infrastructure.

Fixes #8941 and fixes #8964.