Skip descendant dialogs and popovers for autofocus
This patch modifies the "autofocus delegate" algorithm to make it skip
over descendant dialog elements and elements with the popover attribute
when looking for an element to give focus.

The reason behind doing this is that we don't want to focus nested
popovers or dialogs when showing a popover. This used to be implemented
in chromium with an entirely custom built initial focus algorithm for
popovers which didn't use "autofocus delegate", but I got rid of that
custom algorithm and moved this logic into "autofocus delegate" in
chromium here: http://crrev.com/c/4021969

Fixes https://github.com/whatwg/html/issues/9119