Fix form submission's replace-before-load logic
In 524485a38a03bc47478d5cb2e8c1543c654d7870, we overhauled the logic which converts "push" navigations to "replace" navigations before the load event fires. However, for form submissions we erroneously checked the loading state of the form's document, instead of the loading state of the target document. Fix that.

Closes #9135.