Only use the used invoker to establish popover hierarchy
Fixes https://github.com/whatwg/html/issues/9160

If this patch is merged, then https://github.com/whatwg/html/pull/9048
is obsolete and can be closed without merging.

Rationale (copied from [masons patch](https://chromium-review.googlesource.com/c/chromium/src/+/4429412)):

Instead of using just
the DOM structure to establish the popover hierarchy, the user's
behavior should matter. For example, if one popover contains a popover
invoker pointing to another popover, it should matter whether that
invoker is *actually used* to open the second popover.

An example:
- Component 1 is a third party widget, which uses popover
- Component 2 is another third party widget, also using popover
- A page wants to use both components separately, from separate
  invoking buttons.
- Component 1 also wants to be able to use Component 2, via a
  button within Component 1.

In this example, the page should be able to still independently use
these components. So a user clicking the page's button for Component 2
is expected to close Component 1 if it's open, because that's a direct
invocation of Component 2. However, if the user clicks the button
within Component 1 to get Component 2, it is natural to leave Component
1 open because this is a nested call.

Important note: this often happens to be the behavior even before this
change, since the user clicking on the page-level Component 2 invoking
button represents a light dismiss signal for Component 1, so it closes
either way. But this patch simplifies the implementation,
removing the need to track all invokers on the page, and also removing
the need to continuously check whether invoker relationships have
changed.