Fix a corner case where `beforetoggle` opens a new popover
Fixes https://github.com/whatwg/html/issues/9197
Fixes https://github.com/whatwg/html/issues/9196

But there was a case where the `beforetoggle`
event could open a new popover while an old one was closing. This
led to odd situations where asserts and infinite loops ensued.
This PR breaks that loop, or rather converts it to an explicit one,
by repeating the "hide all popovers until" algorithm until all
popovers are closed. That admits an explicit loop, which is the
popover equivalent of `while(true) {}`: when the `beforetoggle`
handler for a popover re-shows the popover.