Account for autofocus attribute for non-modal dialog focusing steps
PR #9121 used the dialog focusing steps with isModal set to true for popovers, so they could honor the autofocus attribute.
However, it is incorrect to consider popovers as modal since they don't create a focus trap, or render the outside content inert.

@scottaohara also recommended that non-modal dialogs focusing steps take in account the autofocus attribute.

So to fix both issues, remove the isModal parameter from the dialog focusing steps algorithm, and account for the autofocus attribute
when set on the dialog itself in all cases (non-modal/modal/popovers).