editorial updates for popover example
removal of the aria-haspopup attribute in the popover example.  It is understandably confusing calling this out and identifying it's purpose, without getting into the weeds of what it is, and how it's not 100% related to the popover attribute.  Instead of getting deep into that here, allude to ARIA having requirements related to the construction of a menu widget.

this PR also contains other minor typo fixes I noticed that I missed in my original PR.