Only restore popover focus if focus is in the popover
Popovers will restore focus to the element which was focused before opening the popover when the popover is closed. However, if a user opens a popover and then tab-navigates away from the popover, focus should not be changed when the popover is closed.

More context: https://chromium-review.googlesource.com/c/chromium/src/+/4322607.

Tests: https://github.com/web-platform-tests/wpt/pull/38922.